+++ ICD.Connect.Partitioning.Commercial/Controls/Occupancy/OccupancyTelemetryNames.cs
@@ -2,7 +2,13 @@
 {
 	public static class OccupancyTelemetryNames
 	{
+		public const string SUPPORTED_FEATURES = "Supported Features";
+		public const string SUPPORTED_FEATURES_CHANGED = "Supported Features Changed";
+
 		public const string OCCUPANCY_STATE = "Occupancy State";
 		public const string OCCUPANCY_STATE_CHANGED = "Occupancy State Changed";
+
+		public const string PEOPLE_COUNT = "People Count";
+		public const string PEOPLE_COUNT_CHANGED = "People Count Changed";
 	}
 }